+++ src/components/layout/NavBar.js
@@ -1,53 +1,40 @@
+import Link from "next/link";
+import Image from "next/image";
 import {
 	ActionIcon,
 	Flex,
 	useMantineColorScheme,
 	Tooltip,
-	Title,
 	Anchor,
 } from "@mantine/core";
-import Link from "next/link";
-import Image from "next/image";
-import { IconBrandTwitter, IconSun, IconMoonStars } from "@tabler/icons";
+import { IconSun, IconMoonStars } from "@tabler/icons";
+import { useHotkeys } from "@mantine/hooks";
 
 const NavBar = () => {
 	const { colorScheme, toggleColorScheme } = useMantineColorScheme();
 	const dark = colorScheme === "dark";
 
+	useHotkeys([["mod+J", () => toggleColorScheme()]]);
+
 	return (
 		<Flex justify="space-between" align="center" w="100%" mb={25}>
-			<Title order={1} fz={20} fw={600}>
-				<Link href="/" passHref legacyBehavior>
-					<Anchor inherit variant="text">
-						<Image
-							alt="Daily place logo"
-							src={dark ? "/logo-dark.svg" : "/logo-light.svg"}
-							width={40}
-							height={40}
-						/>
-					</Anchor>
-				</Link>
-			</Title>
+			<Link href="/" passHref legacyBehavior>
+				<Anchor inherit variant="text">
+					<Image
+						alt="Daily place logo"
+						src={dark ? "/logo-dark.svg" : "/logo-light.svg"}
+						width={40}
+						height={40}
+					/>
+				</Anchor>
+			</Link>
 			<Flex gap="xs">
-				<Tooltip label="Share on twitter">
-					<ActionIcon
-						color="blue"
-						variant="light"
-						aria-label="Share on twitter"
-						className="twitter-share-button"
-						component="a"
-						href="https://twitter.com/intent/tweet?text=daily.place&url=https://www.daily.place&hashtags=lofi,pomodoro,todo"
-						target="_blank"
-					>
-						<IconBrandTwitter size={18} />
-					</ActionIcon>
-				</Tooltip>
-				<Tooltip label="Change theme">
+				<Tooltip label="Toggle theme">
 					<ActionIcon
 						variant="light"
-						color={dark ? "" : "gray"}
+						color={dark ? "brand" : "gray"}
 						onClick={() => toggleColorScheme()}
-						aria-label="Toggle color scheme"
+						aria-label="Toggle theme"
 					>
 						{dark ? (
 							<IconSun size={18} />