+++ src/constants/KeyboardShortcuts.js
@@ -1,0 +1,26 @@
+export const KEYBOARD_SHORTCUTS = [
+	{
+		keys: "<kbd>⌘/ctrl</kbd> + <kbd>J</kbd>",
+		description: "Toggle theme",
+	},
+
+	{
+		keys: "<kbd>⌘/ctrl</kbd> + <kbd>P</kbd>",
+		description: "Play/pause pomodoro timer",
+	},
+
+	{
+		keys: "<kbd>⌘/ctrl</kbd> + <kbd>1</kbd>",
+		description: "Select pomodoro timer",
+	},
+
+	{
+		keys: "<kbd>⌘/ctrl</kbd> + <kbd>2</kbd>",
+		description: "Select short break",
+	},
+
+	{
+		keys: "<kbd>⌘/ctrl</kbd> + <kbd>3</kbd>",
+		description: "Select long break",
+	},
+];