+++ src/constants/Playlist.js
@@ -1,37 +1,42 @@
 export const SOUNDS_LIST = [
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677625605/daily-place/webm/lofi-2_qlbmxk.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/lofi-2.webm",
         videoID: "vkQcpY3T8T4",
         title: "Lo-fi"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677624511/daily-place/webm/coffee-shop-2_crk3xy.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/coffee-shop-2.webm",
         videoID: "h2zkV-l_TbY",
         title: "Coffee shop"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677626304/daily-place/webm/forest-2_al6etg.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/forest-2.webm",
         videoID: "M0AWBnAv8VE",
         title: "Forest"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677624777/daily-place/webm/fireplace-2_yg1zlu.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/fireplace-2.webm",
         videoID: "L_LUpnjgPso",
         title: "Fireplace"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677624878/daily-place/webm/rain-2_otqtyy.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/rain-2.webm",
         videoID: "nDq6TstdEi8",
         title: "Rain"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677626547/daily-place/webm/waves-_fzmlfh.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/waves-.webm",
         videoID: "Nep1qytq9JM",
         title: "Waves"
     },
     {
-        cover: "https://res.cloudinary.com/listi/video/upload/v1677627034/daily-place/webm/vaporwave-2_lcds9j.webm",
+        cover: "https://ik.imagekit.io/dailyplace/covers/vaporwave-2.webm",
         videoID: "zlxXwE0Oop8",
         title: "Vaporwave/Lofi/Synthwave"
+    },,
+    {
+        cover: "https://ik.imagekit.io/dailyplace/covers/asmr-2.webm",
+        videoID: "-SYwOAe6V_4",
+        title: "ASMR"
     },
 ];