+++ src/pages/_app.js
@@ -4,6 +4,7 @@ import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
 import { getCookie, setCookie } from "cookies-next";
 import { MantineProvider, ColorSchemeProvider } from "@mantine/core";
+import { NotificationsProvider } from "@mantine/notifications";
 import splitbee from "@splitbee/web";
 import * as gtag from "../lib/gtag";
 
@@ -76,19 +77,40 @@ export default function App({ Component, pageProps, mode }) {
 					},
 					colors: {
 						brand: [
-							"#f3e7e2",
-							"#f5c7b6",
-							"#f5ab8f",
-							"#f5926d",
-							"#f56e3b",
-							"#f36841",
-							"#f16247",
-							"#ef5c4e",
-							"#eb5358",
-							"#e84a61",
+							"#fff0e7",
+							"#fee0ca",
+							"#fec0a4",
+							"#ffa07b",
+							"#f28a61",
+							"#f56d38",
+							"#f16345",
+							"#ee5851",
+							"#ea4b5f",
+							"#e2394e",
 						],
 					},
 					primaryColor: "brand",
+					components: {
+						Modal: {
+							styles: {
+								title: { fontWeight: 500 },
+							},
+						},
+						TextInput: {
+							styles: {
+								label: {
+									marginBottom: 5,
+								}
+							},
+						},
+						Textarea: {
+							styles: {
+								label: {
+									marginBottom: 5,
+								},
+							},
+						},
+					},
 				}}
 			>
 				<Head>
@@ -159,20 +181,26 @@ export default function App({ Component, pageProps, mode }) {
 						}}
 					/>
 				</Head>
-				<Layout>
-					<style global jsx>{`
-						html,
-						body,
-						div#__next {
-							height: 100%;
-						}
-					`}</style>
-					<Script
-						strategy="afterInteractive"
-						src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
-					/>
-					<Component {...pageProps} />
-				</Layout>
+				<NotificationsProvider
+					position="top-center"
+					color="orange"
+					autoClose={60000}
+				>
+					<Layout>
+						<style global jsx>{`
+							html,
+							body,
+							div#__next {
+								height: 100%;
+							}
+						`}</style>
+						<Script
+							strategy="afterInteractive"
+							src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
+						/>
+						<Component {...pageProps} />
+					</Layout>
+				</NotificationsProvider>
 			</MantineProvider>
 		</ColorSchemeProvider>
 	);
@@ -181,4 +209,3 @@ export default function App({ Component, pageProps, mode }) {
 App.getInitialProps = ({ ctx }) => ({
 	mode: getCookie("daily-place-theme", ctx) || "light",
 });
-